# Making the array