@Cadair 2025-07-15
I've kept chunksize around but made it throw a deprecation warning (which I don't expect people to ever see).